Makes the sweeping gas logic apply only when the gas payer is sending ETH.
This code is a tad sketchy, because it assumes that a `window.ethereum` wallet is the gas payer which isn't necessarily true.
We could perhaps do `eth_getCode` to get the window wallet's code to see if it is a contract.
However, that is an async call and adding an async call will make things quite ugly for a case that may never occur in reality.